refactor: adjust imports and use shared fullscreen functions
