+++ src/components/Item/getAvailableDimensions.js
@@ -1,7 +1,7 @@
 import { isEditMode, isPrintMode } from '../../modules/dashboardModes.js'
 import { getItemHeightPx } from '../../modules/gridUtil.js'
+import { isElementFullscreen } from './fullscreenUtil.js'
 import { getGridItemElement } from './getGridItemElement.js'
-import { isElementFullscreen } from './isElementFullscreen.js'
 import memoizeOne from './memoizeOne.js'
 
 export const getAvailableDimensions = ({
